Arreglar problemas con los cursos y los cargos, crear reporte de inscripciones

+++ src/app/components/cursos/curso/curso.component.ts
@@ -28,14 +28,13 @@ export class CursoComponent implements OnInit {
   
   idparam: string = "";
 
-  niveles: DropDownItem[] = ddNiveles;
+  niveles: DropDownItem[] = [ { name:"Seleccione el nivel", code: "" }, ...ddNiveles];
   nivelSelected : DropDownItem = new DropDownItem();
 
   cargoDialog: boolean = false;
   submitted: boolean = false;
 
   curso: Curso = new Curso();
-  cargos: CargoCurso[] = [];
   cargoSelected: CargoCurso = this.cargoEmpty;
   cargoSelectedIndex: number = -1;
 
@@ -43,31 +42,27 @@ export class CursoComponent implements OnInit {
   searchResultMsg: string = "";
   productoSelected: any = {};
 
-  
-
   constructor(
     public router: Router,
     public activatedRoute: ActivatedRoute,
     private dialogService: DialogService,
-    public cargosService: CargosService,
     public cursosService: CursosService,
     public usuariosService: UsuariosService,
-    private messageService: MessageService
-    // public _modalUploadService: ModalUploadService
+    private messageService: MessageService,
+    private confirmationService: ConfirmationService
   ) {
     activatedRoute.params.subscribe(params => {
-      let id = params.id;
-      this.idparam = id;
-      if (id !== "new") {
-        this.getCurso(id);
+      this.idparam = params.id || "";
+      if (params.id !== "new") {
+        this.getCurso(params.id);
       }
     });
   }
 
   ngOnInit() {
-    this.cargosService
-      .getCargos()
-      .then((resp:any) => (this.cargos = resp.cargos));
+    // this.cargosService
+    //   .getCargos()
+    //   .then((resp:any) => (this.cargos = resp.cargos));
   }
 
   getCurso(id: string) {
@@ -78,11 +73,13 @@ export class CursoComponent implements OnInit {
     this.cursosService.getCursoById(id)
       .then(async (resp) => {      
           const body = await resp.json();
-          console.log(body.curso);
+          if(!body.ok){
+            this.showToastMessage("Error", ` ${ body.msg } ${ body.error || '' }`  , eSeverityMessages.warn );
+            return;
+          }
+          // console.log(body);
           this.curso = body.curso;
-          // this.curso.cargo = curso.cargo.id;
-          // this.cargo = curso.cargo.id;
-          // this.getCargosCurso(this.curso.id);
+          this.nivelSelected = this.getDropDownOption(this.niveles, this.curso.nivel);
     });
   }
 
@@ -139,16 +136,15 @@ export class CursoComponent implements OnInit {
   }
 
   save(){
-    console.log("TODO: Saving...", this.idparam, this.curso);
     if( !this.idparam || this.idparam === "new" ){
       this.cursosService.save(this.curso)
       .then(async(resp)=>{
         const body = await resp.json();
-        console.log(body);
+        console.log("Guardado: ", body);
         if(body.ok){
           this.idparam = body.id;
           this.showToastMessage("Curso", "Guardado con exito", eSeverityMessages.success);
-          this.router.navigate(["/curso", this.curso.id]);
+          this.router.navigate(["/curso/", body.id]);
         } else{
           this.showToastMessage("Curso", "Error al guardar", eSeverityMessages.error);
         }
@@ -159,7 +155,7 @@ export class CursoComponent implements OnInit {
     this.cursosService.update(this.curso)
     .then(async(resp)=>{
       const body = await resp.json();
-      console.log(body);
+      // console.log("Actualizado: ", body);
       if(body.ok){
         this.showToastMessage("Curso", "Actualizado con exito", eSeverityMessages.success);
       }else{
@@ -179,13 +175,20 @@ export class CursoComponent implements OnInit {
   onChangeTipocargo(event: any){
     // console.log(event);
     this.tipocargoSelected = event.value;
-    this.cargoSelected.nombre = event.value.name;
+    if(this.cargoSelected.nombre === ""){
+      this.cargoSelected.nombre = event.value.name;
+    }
     this.productoSearch = event.value.name;
   }
 
-  getTipoCargoOption(searchToken:string, byCode:boolean=false) : any {
-    if (byCode) return this.tiposCargos.find((obj : DropDownItem ) => ( obj.code.toLowerCase() === searchToken.toLowerCase() ));
-    return this.tiposCargos.find((obj : DropDownItem ) => ( obj.name.toLowerCase() === searchToken.toLowerCase() ));
+  // getTipoCargoOption(searchToken:string, byCode:boolean=false) : any {
+  //   if (byCode) return this.tiposCargos.find((obj : DropDownItem ) => ( obj.code.toLowerCase() === searchToken.toLowerCase() ));
+  //   return this.tiposCargos.find((obj : DropDownItem ) => ( obj.name.toLowerCase() === searchToken.toLowerCase() ));
+  // }
+
+  getDropDownOption(dropDownList: DropDownItem[], searchToken:string, byCode:boolean=false) : any {
+    if (byCode) return dropDownList.find((obj : DropDownItem ) => ( obj.code.toLowerCase() === searchToken.toLowerCase() ));
+    return dropDownList.find((obj : DropDownItem ) => ( obj.name.toLowerCase() === searchToken.toLowerCase() ));
   }
 
   showProductosList() {
@@ -213,26 +216,31 @@ export class CursoComponent implements OnInit {
   }
 
   addCharge(){
-    console.log("agregarCargo...", this.productoSelected);
-
     if (!this.validaCargo()) return;
 
     if(!this.productoSelected) return;
 
+    this.cargoSelected.monto = this.cargoSelected.precio * (1 + this.cargoSelected.tasaIVA );
+    this.cargoSelected.tipocargo = this.tipocargoSelected.name;
+
+    // console.log("this.cargoSelected: ", this.cargoSelected);
+    // return;
+    
     if(this.curso.id){
-      // Si el curso ya existe lo guarda de una vez
-      console.log("cargoSelected: ", this.cargoSelected );
-      
+      // Si el curso ya existe lo guarda de una vez      
       if(this.cargoSelected.id){
         // Si el cargo ya existia lo actualizamos 
+        
         this.cursosService.updateCharge(this.cargoSelected)
             .then(async(resp)=>{              
               const body = await resp.json();
-              console.log("respuesta actualizacion: ", body);
+              // console.log("respuesta actualizacion: ", body);
               this.showToastMessage("Cargo", "Actualizado con exito", eSeverityMessages.success);
               this.curso.cargos[this.cargoSelectedIndex] = this.cargoSelected;
             });
       } else {
+        // console.log("Guardar cargo");
+        
         this.cursosService.addCharge({...this.cargoSelected, curso: this.curso.id })
             .then(async(resp)=>{
               const body = await resp.json();
@@ -258,7 +266,7 @@ export class CursoComponent implements OnInit {
     this.cargoSelected = new CargoCurso("", "", "", false);
     this.productoSearch = "";
     this.productoSelected = {};
-    this.tipocargoSelected = this.getTipoCargoOption("", true);
+    this.tipocargoSelected = this.getDropDownOption(this.tiposCargos, "", true);
   }
 
   showToastMessage(title: string = "", text: string = "", tipo: string = "success"){
@@ -288,6 +296,7 @@ export class CursoComponent implements OnInit {
   }
   
   openNew() {
+    this.cleanSelectedCargo();
     this.submitted = false;
     this.cargoDialog = true;
   }
@@ -296,29 +305,38 @@ export class CursoComponent implements OnInit {
     // console.log("Cargo index: ", index);
     this.cargoSelectedIndex = index;
     this.cargoSelected = this.curso.cargos?this.curso.cargos[index]: this.cargoEmpty;
-    this.tipocargoSelected = this.getTipoCargoOption(this.cargoSelected.nombre);
+    this.tipocargoSelected = this.getDropDownOption(this.tiposCargos, this.cargoSelected.tipocargo);
     this.cargoDialog = true;
   }
 
   removeCargo(index: number){
     // console.log(this.curso.cargos[index]);
-    let cargoId = this.curso.cargos[index].id;
-    if (!cargoId){
-      this.curso.cargos =  arrRemoveAt(this.curso.cargos, index);
-      return;
-    }
-    this.cursosService.removeCharge(cargoId)
-      .then(async(resp)=>{
-        const body = await resp.json();
-        // console.log(body);        
-        if(!body.ok){
-          this.showToastMessage("Error", "Error al eliminar el cargo", eSeverityMessages.error);
+
+    this.confirmationService.confirm({
+      message: '¿En verdad deseas eliminar este cargo?',
+      accept: () => {
+
+        let cargoId = this.curso.cargos[index].id;
+        if (!cargoId){
+          this.curso.cargos =  arrRemoveAt(this.curso.cargos, index);
           return;
         }
         
-        this.showToastMessage("Cargo", "Se removio el cargo con exito.", eSeverityMessages.success);
-        this.curso.cargos =  arrRemoveAt(this.curso.cargos, index);
-      });
+        this.cursosService.removeCharge(cargoId)
+          .then(async(resp)=>{
+            const body = await resp.json();
+            // console.log(body);        
+            if(!body.ok){
+              this.showToastMessage("Error", "Error al eliminar el cargo", eSeverityMessages.error);
+              return;
+            }
+            
+            this.showToastMessage("Cargo", "Se removio el cargo.", eSeverityMessages.success);
+            this.curso.cargos =  arrRemoveAt(this.curso.cargos, index);
+          });
+        }
+    });
+
   }
 
   hideDialog(){
